Remove Scheduler and FlushPolicy from core
