+++ core/src/test/scala/filodb.core/columnstore/SegmentSpec.scala
@@ -11,9 +11,9 @@ import org.scalatest.Matchers
 object SegmentSpec {
   implicit val keyHelper = LongKeyHelper(10000L)
 
-  val schema = Seq(Column("first", "dataset1", 0, Column.ColumnType.StringColumn),
-                   Column("last", "dataset1", 0, Column.ColumnType.StringColumn),
-                   Column("age", "dataset1", 0, Column.ColumnType.LongColumn))
+  val schema = Seq(Column("first", "dataset", 0, Column.ColumnType.StringColumn),
+                   Column("last", "dataset", 0, Column.ColumnType.StringColumn),
+                   Column("age", "dataset", 0, Column.ColumnType.LongColumn))
 
   def mapper(rows: Seq[Product]): Iterator[RowReader] = rows.map(TupleRowReader).toIterator
 
@@ -33,6 +33,30 @@ object SegmentSpec {
   def getSortKey(r: RowReader): Long = r.getLong(2)
 
   val firstNames = Seq("Khalil", "Rodney", "Ndamukong", "Terrance", "Peyton", "Jerry")
+
+  // OK, what we want is to test multiple partitions, segments, multiple chunks per segment too.
+  // With default segmentSize of 10000, change chunkSize to say 100.
+  // Thus let's have the following:
+  // "nfc"  0-99  10000-10099 10100-10199  20000-20099 20100-20199 20200-20299
+  // "afc"  the same
+  // 1200 rows total, 6 segments (3 x 2 partitions)
+  // No need to test out of order since that's covered by other things (but we can scramble the rows
+  // just for fun)
+  val schemaWithPartCol = schema ++ Seq(
+    Column("league", "dataset", 0, Column.ColumnType.StringColumn)
+  )
+
+  val largeDataset = dataset.copy(options = Dataset.DefaultOptions.copy(chunkSize = 100),
+                                  partitionColumn = "league")
+
+  val lotLotNames = {
+    for { league <- Seq("nfc", "afc")
+          numChunks <- 0 to 2
+          chunk  <- 0 to numChunks
+          startRowNo = numChunks * 10000 + chunk * 100
+          rowNo  <- startRowNo to (startRowNo + 99) }
+    yield { (names(rowNo % 6)._1, names(rowNo % 6)._2, Some(rowNo.toLong), Some(league)) }
+  }
 }
 
 class SegmentSpec extends FunSpec with Matchers {
